test: 자동차 경주 테스트 코드 추가

+++ src/main/java/racingCar/domain/strategy/RandomMoveStrategy.java
@@ -8,6 +8,6 @@ public class RandomMoveStrategy implements MoveStrategy {
 
     @Override
     public boolean isMovable() {
-        return (int) (RandomNumberGenerator.generateNumber(RANDOM_BOUND)) >= MOVE_CONDITION;
+        return RandomNumberGenerator.generateNumber(RANDOM_BOUND) >= MOVE_CONDITION;
     }
 }
